+++ tools/azsdk-cli/Azure.Sdk.Tools.Cli/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Release History
 
+## 0.5.19 (2026-02-25)
+
+### Bugs Fixed
+
+- Override path extension variable to support running commands on Windows without extension.
+
+## 0.5.18 (2026-02-24)
+
+### Bugs Fixed
+
+- Fix process running issues on Windows when MCP server is used in copilot CLI
+
 ## 0.5.17 (2026-02-18)
 
 ### Features Added